manager/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
black==24.10.0
certifi==2024.12.14
click==8.1.8
elastic-transport==8.17.0
elasticsearch==8.17.1
fastapi==0.115.7
frozenlist==1.5.0
GeoAlchemy2==0.17.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
shapely==2.0.6
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
